Revert "Foundation Classes - TopLoc package update (#849)"
This reverts commit a6e68c7e70a175605fa0779add068143c7bc8ed2.
