+++ src/FoundationClasses/TKMath/GTests/FILES.cmake
@@ -58,5 +58,4 @@ set(OCCT_TKMath_GTests_FILES
   PLib_Test.cxx
   PLib_JacobiPolynomial_Test.cxx
   PLib_HermitJacobi_Test.cxx
-  TopLoc_Location_Test.cxx
 )
