+++ src/FoundationClasses/TKMath/TopLoc/TopLoc_Datum3D.hxx
@@ -44,13 +44,13 @@ public:
   Standard_EXPORT TopLoc_Datum3D(const gp_Trsf& T);
 
   //! Returns a gp_Trsf which, when applied to this datum, produces the default datum.
-  const gp_Trsf& Transformation() const noexcept { return myTrsf; }
+  const gp_Trsf& Transformation() const { return myTrsf; }
 
   //! Returns a gp_Trsf which, when applied to this datum, produces the default datum.
-  const gp_Trsf& Trsf() const noexcept { return myTrsf; }
+  const gp_Trsf& Trsf() const { return myTrsf; }
 
   //! Return transformation form.
-  gp_TrsfForm Form() const noexcept { return myTrsf.Form(); }
+  gp_TrsfForm Form() const { return myTrsf.Form(); }
 
   //! Dumps the content of me into the stream
   Standard_EXPORT void DumpJson(Standard_OStream& theOStream, Standard_Integer theDepth = -1) const;
