+++ src/FoundationClasses/TKMath/TopLoc/TopLoc_Location.lxx
@@ -21,14 +21,14 @@
 
 //=================================================================================================
 
-inline Standard_Boolean TopLoc_Location::IsIdentity() const noexcept
+inline Standard_Boolean TopLoc_Location::IsIdentity() const
 {
   return myItems.IsEmpty();
 }
 
 //=================================================================================================
 
-inline void TopLoc_Location::Identity() noexcept
+inline void TopLoc_Location::Identity()
 {
   myItems.Clear();
 }
@@ -56,10 +56,25 @@ inline const TopLoc_Location& TopLoc_Location::NextLocation() const
 
 //=================================================================================================
 
-inline size_t TopLoc_Location::HashCode() const noexcept
+inline size_t TopLoc_Location::HashCode() const
 {
-  // Use cached hash from the list node
-  return myItems.HashCode();
+  // Hashing base on IsEqual function
+  if (myItems.IsEmpty())
+  {
+    return 0;
+  }
+  size_t                     aHash = opencascade::MurmurHash::optimalSeed<size_t>();
+  TopLoc_SListOfItemLocation items = myItems;
+  size_t                     aCombined[3];
+  while (items.More())
+  {
+    aCombined[0] = std::hash<Handle(TopLoc_Datum3D)>{}(items.Value().myDatum);
+    aCombined[1] = opencascade::hash(items.Value().myPower);
+    aCombined[2] = aHash;
+    aHash        = opencascade::hashBytes(aCombined, sizeof(aCombined));
+    items.Next();
+  }
+  return aHash;
 }
 
 //=================================================================================================
