+++ src/FoundationClasses/TKMath/TopLoc/TopLoc_SListNodeOfItemLocation.hxx
@@ -31,23 +31,11 @@ class TopLoc_SListNodeOfItemLocation : public Standard_Transient
 
 public:
   TopLoc_SListNodeOfItemLocation(const TopLoc_ItemLocation&        I,
-                                 const TopLoc_SListOfItemLocation& aTail,
-                                 size_t                            theHash)
-      : myTail(aTail),
-        myValue(I),
-        myHash(theHash)
-  {
-  }
+                                 const TopLoc_SListOfItemLocation& aTail);
 
-  TopLoc_SListOfItemLocation& Tail() { return myTail; }
+  TopLoc_SListOfItemLocation& Tail() const;
 
-  const TopLoc_SListOfItemLocation& Tail() const { return myTail; }
-
-  TopLoc_ItemLocation& Value() { return myValue; }
-
-  const TopLoc_ItemLocation& Value() const { return myValue; }
-
-  size_t HashCode() const noexcept { return myHash; }
+  TopLoc_ItemLocation& Value() const;
 
   DEFINE_STANDARD_RTTIEXT(TopLoc_SListNodeOfItemLocation, Standard_Transient)
 
@@ -55,7 +43,8 @@ protected:
 private:
   TopLoc_SListOfItemLocation myTail;
   TopLoc_ItemLocation        myValue;
-  size_t                     myHash;
 };
+
+#include <TopLoc_SListNodeOfItemLocation.lxx>
 
 #endif // _TopLoc_SListNodeOfItemLocation_HeaderFile
