+++ src/FoundationClasses/TKernel/GTests/FILES.cmake
@@ -35,5 +35,6 @@ set(OCCT_TKernel_GTests_FILES
   Standard_Handle_Test.cxx
   TCollection_AsciiString_Test.cxx
   TCollection_ExtendedString_Test.cxx
+  TopLoc_Location_Test.cxx
   UnitsAPI_Test.cxx
 )
